backend/requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
sentence-transformers
python-multipart
python-dotenv
numpy
google-genai
openai
